requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
redis==4.5.5
python-dotenv==1.0.0
Werkzeug>=2.3.7
markdown==3.4.1